python/requirements.txt
bleach==6.1.0
bokeh==3.6.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
idna==3.10
Jinja2==3.1.4
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
panel==1.5.2
param==2.1.1
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyusb==1.2.1
pyviz_comms==3.0.3
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
webencodings==0.5.1
xyzservices==2024.9.0
